call transaction

+++ src/main/kotlin/io/openfuture/api/repository/Repositories.kt
@@ -29,7 +29,7 @@ interface ScaffoldRepository : BaseRepository<Scaffold> {
 
     fun findAllByOpenKeyUser(user: User, pageable: Pageable): Page<Scaffold>
 
-    fun countByEnabledIsFalseAndOpenKeyUser(user: User): Long
+    fun countByEnabledIsFalseAndOpenKeyUser(user: User): Int
 
 }
